implemented skeleton loading and added clean up
